backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1
chromadb==0.4.24
numpy==1.26.4
groq==0.9.0
python-multipart==0.0.9
